feature: displayUserDetails function added
