+++ assets/js/inc/user.js
@@ -1,14 +1,23 @@
 // Responsável por renderizar as informações no DOM
 
-const userTags = {
-  avatar_url: '#userAvatar',
-  bio: '#userBio',
-  location: '#userLocation',
-  name: '#userName',
-  html_url: '#userUrl',
-  followers: '#userFollowersCount',
-  following: '#userFollowingCount',
-  public_repos: '#userReposCount'
+import { selector } from '../helpers/general.js'
+import userTags from '../templates/userTags.js'
+
+const displayUserDetails = details => {
+  for (let userTag in userTags) {
+    const prop = selector(userTags[userTag]),
+      value = details[userTag]
+    switch (userTag) {
+      case 'avatar_url':
+        prop.setAttribute('src', value)
+        break
+      case 'html_url':
+        prop.setAttribute('href', value)
+        break
+      default:
+        prop.innerText = value
+    }
+  }
 }
 
-export default userTags
+export { displayUserDetails }
