+++ assets/js/templates/userTags.js
@@ -1,1 +1,14 @@
 // Responsável por mapear os IDs dos elementos referentes ao usuário no DOM
+
+const userTags = {
+  avatar_url: '#userAvatar',
+  bio: '#userBio',
+  location: '#userLocation',
+  name: '#userName',
+  html_url: '#userUrl',
+  followers: '#userFollowersCount',
+  following: '#userFollowingCount',
+  public_repos: '#userReposCount'
+}
+
+export default userTags
